Move standard models to the standard lib
