+++ docs/source/api_ref/index.rst
@@ -11,5 +11,4 @@ Enaml API Reference
     Parsing <enaml.parsing>
     Styling <enaml.styling>
     Widgets <enaml.widgets>
-    Item Models <enaml.item_models>
     Utility <enaml.util>
