--- enaml/item_models/__init__.py
@@ -1,4 +1,0 @@
-#------------------------------------------------------------------------------
-#  Copyright (c) 2011, Enthought, Inc.
-#  All rights reserved.
-#------------------------------------------------------------------------------
