+++ examples/components/simple_tables.enaml
@@ -8,7 +8,8 @@ An example which shows some simple instances of ListModel and TableModel
 as well as how both can be used in a TableView.
 
 """
-from enaml.item_models.standard_models import ListModel, TableModel
+from enaml.stdlib.list_model import ListModel
+from enaml.stdlib.table_model import TableModel
 from enaml.styling.brush import Brush
 from enaml.styling.color import Color
